Open markdown links in new tab thanks to @drmuey
This is fixed due to @drmuey's https://github.com/chjj/marked/pull/659
which has not been merged with marked.js. I minified this change and
pasted it into the minified version of the file originally included with
the PR to Scrumblr, so only God knows what version of marked.js this now
is.